Refactored max blur pooling
